@@ -24,9 +24,9 @@
     },
 
     "require": {
-        "php": "^5.6|^7.0|^8.0",
+        "php": "^7.3|^8.0",
         "pimple/pimple": "^3.0",
-        "twig/twig": "^1.0|^2.0"
+        "twig/twig": "^2.0|^3.0"
     },
 
     "require-dev": {
